move decorateWithReadConcern to utils
