+++ lib/operations/collection_ops.js
@@ -5,6 +5,8 @@ const checkCollectionName = require('../utils').checkCollectionName;
 const Code = require('mongodb-core').BSON.Code;
 const createIndexDb = require('./db_ops').createIndex;
 const decorateCommand = require('../utils').decorateCommand;
+const decorateWithCollation = require('../utils').decorateWithCollation;
+const decorateWithReadConcern = require('../utils').decorateWithReadConcern;
 const ensureIndexDb = require('./db_ops').ensureIndex;
 const evaluate = require('./db_ops').evaluate;
 const executeCommand = require('./db_ops').executeCommand;
@@ -298,36 +300,6 @@ function createIndexes(coll, indexSpecs, options, callback) {
     options,
     callback
   );
-}
-
-function decorateWithCollation(command, coll, options) {
-  // Do we support collation 3.4 and higher
-  const capabilities = coll.s.topology.capabilities();
-  // Do we support write concerns 3.4 and higher
-  if (capabilities && capabilities.commandsTakeCollation) {
-    if (options.collation && typeof options.collation === 'object') {
-      command.collation = options.collation;
-    }
-  }
-}
-
-function decorateWithReadConcern(command, coll, options) {
-  let readConcern = Object.assign({}, command.readConcern || {});
-  if (coll.s.readConcern) {
-    Object.assign(readConcern, coll.s.readConcern);
-  }
-
-  if (
-    options.session &&
-    options.session.supports.causalConsistency &&
-    options.session.operationTime
-  ) {
-    Object.assign(readConcern, { afterClusterTime: options.session.operationTime });
-  }
-
-  if (Object.keys(readConcern).length > 0) {
-    Object.assign(command, { readConcern: readConcern });
-  }
 }
 
 /**
@@ -1427,8 +1399,6 @@ module.exports = {
   countDocuments,
   createIndex,
   createIndexes,
-  decorateWithCollation,
-  decorateWithReadConcern,
   deleteMany,
   deleteOne,
   distinct,
