+++ lib/utils.js
@@ -586,6 +586,25 @@ function decorateWithCollation(command, target, options) {
   }
 }
 
+function decorateWithReadConcern(command, coll, options) {
+  let readConcern = Object.assign({}, command.readConcern || {});
+  if (coll.s.readConcern) {
+    Object.assign(readConcern, coll.s.readConcern);
+  }
+
+  if (
+    options.session &&
+    options.session.supports.causalConsistency &&
+    options.session.operationTime
+  ) {
+    Object.assign(readConcern, { afterClusterTime: options.session.operationTime });
+  }
+
+  if (Object.keys(readConcern).length > 0) {
+    Object.assign(command, { readConcern: readConcern });
+  }
+}
+
 module.exports = {
   filterOptions,
   mergeOptions,
@@ -608,6 +627,7 @@ module.exports = {
   applyWriteConcern,
   convertReadPreference,
   isPromiseLike,
+  getReadPreference,
   decorateWithCollation,
-  getReadPreference
+  decorateWithReadConcern
 };
